Add parameter vm_identify_interfaces_by_parent
Now by default the network interfaces will be identified by the order
in which they are defined and not by the parent interface, thus allowing
to have multiple network interfaces enslaved to the same master.

If vm_identify_interfaces_by_parent is set to true then the old approach is
used.